OMG SO BEAUTIFUL! Thank you.
I've got a couple of changes I'd like to make. Let's start with the settings page.

Let's move the API Access information to it's own panel and be at the very top.

Let's then change the Settings Panel potentially support multiple webhooks. Each webhook that is set up will require the URL and then there should be checkboxes for charging status and screen orientation. The default webhook should have both of these checked. At the bottom of the section, there should be a button to create another webhook trigger.