I got feedback from users. They'd like the settings accessed through a button on the top right, not as a bottom navigation button. That would get rid of the bottom navigation entirely.
I think if that's the case, then the main page should have the explanation of the app and then either a list of a set of cards that display the webhooks that are set up.

There should be a button on the bottom right to create a new trigger type. The device api statuses should also remain separate.